requirements.txt
Flask
mysql-connector-python
twilio
python-dotenv
schedule
